pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Gabriel Martin del Campo y Eternod", email = "gmartindelcye@gmail.com"},
]
dependencies = [
    "fastapi>=0.104.1",
    "sqlmodel>=0.0.12",
    "asyncpg>=0.29.0",
    "alembic>=1.12.1",
    "httpx>=0.25.2",
    "python-dotenv>=1.0.0",
    "uvicorn>=0.24.0.post1",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "python-jose[criptography]>=3.3.0",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

[tool.pdm.dev-dependencies]
lint = [
    "flake8>=6.1.0",
    "black>=23.11.0",
]
test = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
]
